Fixed Order of GL Context Creation and some warnings.

+++ include/inviwo/core/inviwoapplication.h
@@ -66,7 +66,7 @@ private:
 
     bool initialized_;
 
-    uint8_t* allocTest_;
+    uint32_t* allocTest_;
 
     static const std::string logSource_; ///< Source string to be displayed for log messages.
 };
